# perf: optimize Error() method and add comprehensive testing